Adding tests for Table#getFullName
Some basic tests for the concatenation of the full name of a table
consisting of db, schema and catalog names, depending whether they are
null. Leaving the implementation logic as is.
I doubt the actual implementation is intended to support some of the
cases so I marked them with TODOs.
(Additionally we should consider to make this method non static. But
that should be a separate commit.)