Introduce AssertJ as test dependency, remove JUnit-addons, Refactoring
In preparation to extend the test coverage of the project this commit
introduces AssertJ as a support to write tests in a fluent way.

As an example I replaced the JUnit-addons dependency and the usage of
its ListAssert method in the DBAnalyzerTest. I combined the two
assertions where the first one checked the list for emptyness, as this is
already included in the second assertion (containsExactly...).

I also removed the inheritance of TestCase as it is obsolete since JUnit
4.